Move SummarizedActionState into action_diff.rs
