+++ src/systems.rs
@@ -1,6 +1,5 @@
 //! The systems that power each [`InputManagerPlugin`](crate::plugin::InputManagerPlugin).
 
-use crate::action_state::ActionKindData;
 use crate::prelude::updating::CentralInputStore;
 use crate::user_input::{AccumulatedMouseMovement, AccumulatedMouseScroll};
 use crate::{
@@ -9,15 +8,13 @@ use crate::{
 
 use bevy::ecs::prelude::*;
 use bevy::prelude::Gamepads;
-use bevy::utils::HashSet;
 use bevy::{
     input::mouse::{MouseMotion, MouseWheel},
-    math::Vec2,
     time::{Real, Time},
-    utils::{HashMap, Instant},
+    utils::Instant,
 };
 
-use crate::action_diff::{ActionDiff, ActionDiffEvent};
+use crate::action_diff::{ActionDiffEvent, SummarizedActionState};
 
 /// We are about to enter the `Main` schedule, so we:
 /// - save all the changes applied to `state` into the `fixed_update_state`
@@ -172,266 +169,6 @@ pub fn generate_action_diffs<A: Actionlike>(
         SummarizedActionState::summarize(global_action_state, action_state_query);
     current_action_state.send_diffs(&previous_action_state, &mut action_diff_events);
     *previous_action_state = current_action_state;
-}
-
-/// Stores the state of all actions in the current frame.
-///
-/// Inside of the hashmap, [`Entity::PLACEHOLDER`] represents the global / resource state of the action.
-#[derive(Debug)]
-pub struct SummarizedActionState<A: Actionlike> {
-    button_state_map: HashMap<Entity, HashMap<A, bool>>,
-    axis_state_map: HashMap<Entity, HashMap<A, f32>>,
-    dual_axis_state_map: HashMap<Entity, HashMap<A, Vec2>>,
-}
-
-impl<A: Actionlike> SummarizedActionState<A> {
-    /// Returns a list of all entities that are contained within this data structure.
-    ///
-    /// This includes the global / resource state, using [`Entity::PLACEHOLDER`].
-    pub fn all_entities(&self) -> HashSet<Entity> {
-        let mut entities = HashSet::new();
-        let button_entities = self.button_state_map.keys();
-        let axis_entities = self.axis_state_map.keys();
-        let dual_axis_entities = self.dual_axis_state_map.keys();
-
-        entities.extend(button_entities);
-        entities.extend(axis_entities);
-        entities.extend(dual_axis_entities);
-
-        entities
-    }
-
-    /// Captures the raw values for each action in the current frame.
-    pub fn summarize(
-        global_action_state: Option<Res<ActionState<A>>>,
-        action_state_query: Query<(Entity, &ActionState<A>)>,
-    ) -> Self {
-        let mut button_state_map = HashMap::default();
-        let mut axis_state_map = HashMap::default();
-        let mut dual_axis_state_map = HashMap::default();
-
-        if let Some(global_action_state) = global_action_state {
-            let mut per_entity_button_state = HashMap::default();
-            let mut per_entity_axis_state = HashMap::default();
-            let mut per_entity_dual_axis_state = HashMap::default();
-
-            for (action, action_data) in global_action_state.all_action_data() {
-                match &action_data.kind_data {
-                    ActionKindData::Button(button_data) => {
-                        per_entity_button_state.insert(action.clone(), button_data.pressed());
-                    }
-                    ActionKindData::Axis(axis_data) => {
-                        per_entity_axis_state.insert(action.clone(), axis_data.value);
-                    }
-                    ActionKindData::DualAxis(dual_axis_data) => {
-                        per_entity_dual_axis_state.insert(action.clone(), dual_axis_data.pair);
-                    }
-                }
-            }
-
-            button_state_map.insert(Entity::PLACEHOLDER, per_entity_button_state);
-            axis_state_map.insert(Entity::PLACEHOLDER, per_entity_axis_state);
-            dual_axis_state_map.insert(Entity::PLACEHOLDER, per_entity_dual_axis_state);
-        }
-
-        for (entity, action_state) in action_state_query.iter() {
-            let mut per_entity_button_state = HashMap::default();
-            let mut per_entity_axis_state = HashMap::default();
-            let mut per_entity_dual_axis_state = HashMap::default();
-
-            for (action, action_data) in action_state.all_action_data() {
-                match &action_data.kind_data {
-                    ActionKindData::Button(button_data) => {
-                        per_entity_button_state.insert(action.clone(), button_data.pressed());
-                    }
-                    ActionKindData::Axis(axis_data) => {
-                        per_entity_axis_state.insert(action.clone(), axis_data.value);
-                    }
-                    ActionKindData::DualAxis(dual_axis_data) => {
-                        per_entity_dual_axis_state.insert(action.clone(), dual_axis_data.pair);
-                    }
-                }
-            }
-
-            button_state_map.insert(entity, per_entity_button_state);
-            axis_state_map.insert(entity, per_entity_axis_state);
-            dual_axis_state_map.insert(entity, per_entity_dual_axis_state);
-        }
-
-        Self {
-            button_state_map,
-            axis_state_map,
-            dual_axis_state_map,
-        }
-    }
-
-    /// Generates an [`ActionDiff`] for button data,
-    /// if the button has changed state.
-    ///
-    ///
-    /// Previous values will be treated as default if they were not present.
-    pub fn button_diff(
-        action: A,
-        previous_button: Option<bool>,
-        current_button: Option<bool>,
-    ) -> Option<ActionDiff<A>> {
-        let previous_button = previous_button.unwrap_or_default();
-        let current_button = current_button?;
-
-        if previous_button != current_button {
-            if current_button {
-                Some(ActionDiff::Pressed { action })
-            } else {
-                Some(ActionDiff::Released { action })
-            }
-        } else {
-            None
-        }
-    }
-
-    /// Generates an [`ActionDiff`] for axis data,
-    /// if the axis has changed state.
-    ///
-    /// Previous values will be treated as default if they were not present.
-    pub fn axis_diff(
-        action: A,
-        previous_axis: Option<f32>,
-        current_axis: Option<f32>,
-    ) -> Option<ActionDiff<A>> {
-        let previous_axis = previous_axis.unwrap_or_default();
-        let current_axis = current_axis?;
-
-        if previous_axis != current_axis {
-            Some(ActionDiff::AxisChanged {
-                action,
-                value: current_axis,
-            })
-        } else {
-            None
-        }
-    }
-
-    /// Generates an [`ActionDiff`] for dual axis data,
-    /// if the dual axis has changed state.
-    pub fn dual_axis_diff(
-        action: A,
-        previous_dual_axis: Option<Vec2>,
-        current_dual_axis: Option<Vec2>,
-    ) -> Option<ActionDiff<A>> {
-        let previous_dual_axis = previous_dual_axis.unwrap_or_default();
-        let current_dual_axis = current_dual_axis?;
-
-        if previous_dual_axis != current_dual_axis {
-            Some(ActionDiff::DualAxisChanged {
-                action,
-                axis_pair: current_dual_axis,
-            })
-        } else {
-            None
-        }
-    }
-
-    /// Generates all [`ActionDiff`]s for a single entity.
-    pub fn entity_diffs(
-        &self,
-        previous_button_state: Option<&HashMap<A, bool>>,
-        current_button_state: Option<&HashMap<A, bool>>,
-        previous_axis_state: Option<&HashMap<A, f32>>,
-        current_axis_state: Option<&HashMap<A, f32>>,
-        previous_dual_axis_state: Option<&HashMap<A, Vec2>>,
-        current_dual_axis_state: Option<&HashMap<A, Vec2>>,
-    ) -> Vec<ActionDiff<A>> {
-        let mut action_diffs = Vec::new();
-
-        if let Some(current_button_state) = current_button_state {
-            for (action, current_button) in current_button_state {
-                let previous_button = previous_button_state
-                    .and_then(|previous_button_state| previous_button_state.get(action))
-                    .copied();
-
-                if let Some(diff) =
-                    Self::button_diff(action.clone(), previous_button, Some(*current_button))
-                {
-                    action_diffs.push(diff);
-                }
-            }
-        }
-
-        if let Some(current_axis_state) = current_axis_state {
-            for (action, current_axis) in current_axis_state {
-                let previous_axis = previous_axis_state
-                    .and_then(|previous_axis_state| previous_axis_state.get(action))
-                    .copied();
-
-                if let Some(diff) =
-                    Self::axis_diff(action.clone(), previous_axis, Some(*current_axis))
-                {
-                    action_diffs.push(diff);
-                }
-            }
-        }
-
-        if let Some(current_dual_axis_state) = current_dual_axis_state {
-            for (action, current_dual_axis) in current_dual_axis_state {
-                let previous_dual_axis = previous_dual_axis_state
-                    .and_then(|previous_dual_axis_state| previous_dual_axis_state.get(action))
-                    .copied();
-
-                if let Some(diff) = Self::dual_axis_diff(
-                    action.clone(),
-                    previous_dual_axis,
-                    Some(*current_dual_axis),
-                ) {
-                    action_diffs.push(diff);
-                }
-            }
-        }
-
-        action_diffs
-    }
-
-    /// Compares the current frame to the previous frame, generates [`ActionDiff`]s and then sends them as batched [`ActionDiffEvent`]s.
-    pub fn send_diffs(&self, previous: &Self, writer: &mut EventWriter<ActionDiffEvent<A>>) {
-        for entity in self.all_entities() {
-            let owner = if entity == Entity::PLACEHOLDER {
-                None
-            } else {
-                Some(entity)
-            };
-
-            let previous_button_state = previous.button_state_map.get(&entity);
-            let current_button_state = self.button_state_map.get(&entity);
-            let previous_axis_state = previous.axis_state_map.get(&entity);
-            let current_axis_state = self.axis_state_map.get(&entity);
-            let previous_dual_axis_state = previous.dual_axis_state_map.get(&entity);
-            let current_dual_axis_state = self.dual_axis_state_map.get(&entity);
-
-            let action_diffs = self.entity_diffs(
-                previous_button_state,
-                current_button_state,
-                previous_axis_state,
-                current_axis_state,
-                previous_dual_axis_state,
-                current_dual_axis_state,
-            );
-
-            writer.send(ActionDiffEvent {
-                owner,
-                action_diffs,
-            });
-        }
-    }
-}
-
-// Manual impl due to A not being bounded by Default messing with the derive
-impl<A: Actionlike> Default for SummarizedActionState<A> {
-    fn default() -> Self {
-        Self {
-            button_state_map: Default::default(),
-            axis_state_map: Default::default(),
-            dual_axis_state_map: Default::default(),
-        }
-    }
 }
 
 /// Release all inputs when an [`InputMap<A>`] is removed to prevent them from being held forever.
